Write constructors and initializers for the tubserv_t and tubcli_t types.
Create the constructors, and provide tests for each of them to make sure they work.
The tests provided in *test.c files should be able to connect to one another.

Stop calling htons() on the port; it should be host order when converted to a string
so getaddrinfo can handle it.