This commit strips away miknet's object implementations.
In future commits on this branch I will be implementing mikserv and mikcli
as one object, miknode. This will make sure the library does not require
programs using it to have a centralized model, which would be quite the
obstacle!

Stripping it all away at once currently makes for a convoluted commit, but
it means that the actual implementation can have meaningful ones, which I
think is more important.