Create a constructor and deconstructor for the peer structure.
Peers are created over the SOCK_STREAM socket, and so that is preferred.
The array which holds all the peers of a node, node.peers, determines whether
a slot is empty or occupied by the state of the peer. MIK_CONN is active, and
MIK_DISC is an empty slot.