Remove any handling for datagram sockets. They will not be included.
Rationale:

TCP alongside UDP will only work in a few ways, all of which compromise
the usefulness of the library. A secret exchange could be used to associate
the two sockets with other peers, but it remains vulnerable to MITM if the
data following connection is in plaintext.

A secret exchange followed by encrypted traffic defeats the usefulness of
having a MIK_FAST packet flag.

Associating the two sockets by having each session share a single port
number only works on IPV4 if the machines are not behind NAT, which would
eliminate most users. On the bright side, this kind of thing will be
plausible upon the widespread adoption of IPV6!

The original purpose of the library was for chat applications, for which
UDP is not particularly relevant. I wanted to include it in case the
library was used for other things (like games), but it isn't robust
enough for that anyway, and I think it is better to have a good tool for
one use over a mediocre tool for two uses.