Change file descriptor array offset to 1, and clean up after peers.
The file descriptor array offset was previously two, because aside from
peers, each node had a udp socket, and a master tcp socket. Now there is
only a master tcp socket preceding the sockets for peers.

After a peer is done with a slot, the file descriptor should be set back to
-1 so it is ignored in poll.