Implement sending and receiving messages.
Miknet can now send and receive messages. mikpeer_send will
queue a packet to be sent, and it will send when the miknode is allowed
to operate (poll).

Some small fixes are also included, such as a typo left over offsetting the
fds array by 2 instead of 1. Nothing that should change the intended
behavior or rationale behind those modules.

The test code is updated for convenience.