Replace mikevent_t type with miklist_t type.
This was something I intened to earlier. Since miklists only store mikevent
types, there is no reason for them to have a seperate tier, with a pointer
to the data and adjustable length. It just makes things confusing.

Now mikevents no longer exist, and it is all contains in miklist_t. And as
an unintended positive externality, all the memory leaks are fixed!