Complete the transition to vectors.
The entire library now respects the queues as vectors, not linked lists. In
order to make things easier (as, admittedly, linked lists were easier to use as
as a programmer using the library) a function mikevent was added in miknet.c,
which makes it feel the same.

All the changes have been desk-checked and against valgrind; leak-free (for
now)!