Reduce miknet header and make it network-safe.
Some of the data in mikpack_t (such as the pointer to the data, and the
peer index) were useless to machines other than the sender. They also
varied in width and byte order.

mikmeta_t is strictly the network-relevant metadata for a packet. Functions
have been implemented to encode it in network byte other and interpret it
back to host byte order. The types in it have explicitly defined bit widths
and altogether amount to 7 octets for the header, which isn't bad at all.

On all machines, it should be 7 octets. However, on machines that do not
define bytes as 8 bits there could be some problems.