Rewrites miksys.
miksys should be more general than just the posix function wrappers it provides.
In the future it also needs to wrap memory functions.

This changes the naming of the components in miksys to communicate that the
posix function wrapper is component of it, and not the whole of it.

Other changes have been made to provide referential transparency for the
functions which may use miksys wrappers. Instead of calling miksys functions,
they will make calls from a wrapper they are provided.

I hope to in the future find a nice way to associate mock behaviors with a
wrapper, rather than requiring miksysmock to retain a state.