Grants ref transparency in mikaddr, removes con.
Testing with miksysmock prior to this commit could yield different behavior by
calling the functions in mikaddr with the same output, based on some state
maintained in miksysmock.c. This commit removes the need for global state, and
also allows some flexibility in configuring the behaviors of the mocked
functions.

As an example, two functions can now be used and have their mock behavior
independently controlled, rather than having one mock return value for all
functions called between uses of a mock.

Connect is also removed as behavior from mikaddr. There will be no more
connecting because miknet is going udp! This will be addressed in a future
commit more thoroughly, but it made more sense to go ahead and remove that than
update the code I intended to remove next commit.