Changes miknet basic to udp.
The basic beneath miknet is now udp, unlike the Mercury release. Really, it's
weird that it's just a different version of the same project.

mikaddr should be passive so that mikpeer can use the information it holds. The
construction function is just convenient because of how many ways creating this
information can fail; it should not imply that mikaddr has behaviors.
mikaddr_close has been removed to avoid this implication.