Implements queue structure.
This structure will hold packets on a per-peer per-channel basis. As well as
peers for the node and whatever comes up. In the future, there may need to be
some revisiting to give thought to more/larger contiguous allocations.

Purposes for the miklist should have syntactic sugar added to avoid typecasting
nonsense no one can read. Derefencing void pointers is bad for the eyes.