Removes channel information from packets.
This commit removes channel information from packets, and adds a packet type
which indicates that the packet is sent. In essence, its type is only a
receipt, not an outgoing command. This will be used to keep track of what
fragments in a packet have been acknowledged by a peer.

Some variables have been renamed and reorganized for readability.