Reorganizes cmake files for check.
There are some problems with the standard FindCheck module on Ubuntu. I didn't
realize this because until now I had been working on Debian.

On Ubuntu pthread isn't included in the pkg-config's return from cmake, meaning
the cflags also have to be specified (where the pthread link is). Oddly, pthread
shows up if "pkg-config check --libs" is invoked in bash.