Creates mikmsg type.
mikmsg_t object represent a higher level unit of transmitting data, composed of
one or more mikgrams. When users receive messages it will be available to them
in the form of a mikmsg_t, rather than a mikgram_t, because they could be
receiving a message sent using many mikgrams.